02: Arcjet Env Type Safety Setup
